requirements.txt
discord.py
google-api-python-client
flask